cli: move publishing utils to maintenance module
