+++ packages/cli/src/modules/maintenance/commands/package/pack.ts
@@ -20,7 +20,7 @@ import {
 } from '../../../../modules/build/lib/packager/productionPack';
 import { paths } from '../../../../lib/paths';
 import fs from 'fs-extra';
-import { publishPreflightCheck } from '../../../../lib/publishing';
+import { publishPreflightCheck } from '../../lib/publishing';
 import { createTypeDistProject } from '../../../../lib/typeDistProject';
 
 export const pre = async () => {
